Compile contacts plugin metadata in the actual QML plugin, not in the shared library.
Otherwise Qt refuses to load it. This is worked around currently by linking the binaries to the contacts shared library, but this does not work when linking with --as-needed (as most distros do nowadays).

Also disambiguate the target names, currently both the shared library and the QML plugin have the same cmake target name.

BUG: 491808