mail: Fix MailComposer not showing the right sender email address
The `identity` ComboBox used to refresh the sender email onCurrentIndexChanged,
but this signal seems to be sent *before* the currentValue of the ComboBox is
changed. onActivated is ran *after*, which has the expected behaviour.

This patch also makes it use `Component.onCompleted` to set an email when the
model is loaded